extracted the code trying to determine the encoding of the input file to a separate module
